requirements.txt
blend-modes==2.1.0
pandas==1.4.2
Pillow==9.1.1
requests==2.27.1
